v148-modified blogServices.js

+++ back-end/src/services/blogServices.js
@@ -1,3 +1,4 @@
+//note: used serviceResponse but I didnt checked it works properly or not
 const mongoose = require("mongoose");
 const deleteFile = require("../helpers/deleteFile");
 const Blog = require("../models/blog");
@@ -6,18 +7,20 @@ const serviceResponse = require("../helpers/serviceResponse");
 class BlogServices {
   async getAllBlogs(req, res) {
     //خواندن تمام مقالات از دیتابیس
-    return Blog.find({});
+    const findOp = await Blog.find({});
+    return serviceResponse(200, findOp);
   }
   async seeOneBlog(req, res) {
     // خواندن یک مقاله از دیتابیس
-    return Blog.findById(req.params.blogId);
+    const findOp = await Blog.findById(req.params.blogId);
+    return serviceResponse(200, findOp);
   }
 
   //note:this function needs to fix .
   async getBlogsByCategoryString(string, req, res) {
     //خواندن مقالات مخصوص دسته بندی انتخاب شده از دیتابیس
     let array = [];
-    const blogs = await this.getAllBlogs(req, res);
+    const { data: blogs } = await this.getAllBlogs(req, res);
     if (blogs) {
       blogs.forEach((blog) => {
         if (blog.categoryId && string.includes(blog.categoryId.toString())) {
@@ -25,7 +28,7 @@ class BlogServices {
         }
       });
     }
-    return array;
+    return serviceResponse(200, array);
   }
 
   async createBlog(req, res) {
@@ -41,33 +44,30 @@ class BlogServices {
     if (req.file) {
       newBlog.img = "/" + req.file.path.replace(/\\/g, "/"); //تنظیم آدرس تصویر مقاله برای ذخیره در مونگو دی بی
     }
-    return newBlog.save();
+    const saveOp = await newBlog.save();
+    return serviceResponse(200, saveOp);
   }
 
   //بروزرسانی مقاله
   async updateBlog(req, res) {
-    const blog = await this.seeOneBlog(req, res);
-    let data = {
-      title: req.body.title,
-      author: req.body.author,
-      categoryId: blog.categoryId,
-      description: req.body.description,
-      img: blog.img,
-    };
-    if (req.body.categoryId) {
-      data.categoryId = new mongoose.Types.ObjectId(req.body.categoryId);
+    const { data: blog } = await this.seeOneBlog(req, res);
+    if (blog) {
+      blog.title = req.body.title;
+      blog.author = req.body.author;
+      blog.description = req.body.description;
+      if (req.body.categoryId) {
+        blog.categoryId = new mongoose.Types.ObjectId(req.body.categoryId);
+      }
+      if (req.file) {
+        blog.img
+          ? deleteFile(blog.img.substring(1), blog.img.substring(1))
+          : null;
+        blog.img = "/" + req.file.path.replace(/\\/g, "/"); //تنظیم آدرس تصویر پروفایل برای ذخیره در مونگو دی بی
+      }
+      const updateOp = await blog.save();
+      return serviceResponse(200, {});
     }
-    if (req.file) {
-      blog.img
-        ? deleteFile(blog.img.substring(1), blog.img.substring(1))
-        : null;
-      data.img = "/" + req.file.path.replace(/\\/g, "/"); //تنظیم آدرس تصویر پروفایل برای ذخیره در مونگو دی بی
-    }
-    const updateOp = await Blog.updateOne({ _id: blog.id }, { $set: data });
-    if (updateOp.modifiedCount.valueOf() > 0) {
-      return true;
-    }
-    return false;
+    return serviceResponse(404, {});
   }
 
   //انتقال مقالات به کتگوری دیگر بعد از حذف کتگوری فعلی
@@ -85,17 +85,14 @@ class BlogServices {
 
   async deleteBlog(req, res) {
     //حذف مقاله
-    const blog = await this.seeOneBlog(req, res);
-    if (blog) {
-      blog.img
-        ? deleteFile(blog.img.substring(1), blog.img.substring(1))
+    const deleteOp = await Blog.findOneAndDelete({ _id: req.params.blogId });
+    if (deleteOp) {
+      deleteOp.img
+        ? deleteFile(deleteOp.img.substring(1), deleteOp.img.substring(1))
         : null;
-      const deleteOp = await Blog.deleteOne({ _id: req.params.blogId });
-      if (deleteOp.deletedCount.valueOf() > 0) {
-        return true;
-      }
+      return serviceResponse(200, {});
     }
-    return false;
+    return serviceResponse(404, {});
   }
 }
 module.exports = new BlogServices();
